// src/types.ts
export interface LocationRad {
  longitudeRad: number;
  latitudeRad: number;
}

export interface HorizonSkyOptions {
  color?: string;
  opacity?: number;
}

export interface HorizontalRad {
  altRad: number;
  azRad: number;
}

export interface EquatorialRad {
  raRad: number;
  decRad: number;
}

export interface LocationDeg {
  longitudeDeg: number;
  latitudeDeg: number;
}

export interface AltAzRad {
  altRad: number;
  azRad: number;
}